requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
pymilvus>=2.3.3
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
spacy>=3.0.0
unstructured>=0.10.0
langdetect>=1.0.0
crewai>=0.1.0
ollama>=0.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Optional LLM providers
anthropic>=0.7.0  # For Claude
openai>=1.0.0  # For GPT models

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0  # For FastAPI testing 